src/requirements.txt
mitmproxy
selenium
pymongo
sqlalchemy 
pandas
python-dotenv
requests[socks]